pyproject.toml
[tool.poetry]
name = "geocode-converter"
version = "0.1.0"
description = ""
authors = ["Luis Carlos <luiscarlosfirminopinheiro@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = {extras = ["speedups"], version = "^3.6.2"}
python-dotenv = "^0.14.0"
ratelimit = "^2.2.1"
openpyxl = "^3.0.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
pylint = "^2.6.0"
mypy = "^0.782"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
